Extract Token definition into token.rs
